+++ tools/lossy-up.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -o errexit
+
+exec tc qdisc add dev lo root netem loss 5% 25%